requirements.txt
requests
pyopenssl
pyyaml